Decouple js/package.json version from setuptools-scm
`make build` regenerated the version from git tags, so the tracked value
could never match what setuptools-scm computes once a release tag exists:
the tag lands on a commit whose package.json predates it. Every commit
after a release failed `prek run --all-files` — after v0.2.18 the
rebuild-from-specs hook rewrote 0.2.18 to 0.2.19 and reported a diff.

Version writing moves to scripts/set_npm_version.py, run only by
`make release-npm`; the tracked value is a 0.0.0 placeholder. It now
rejects non-release versions instead of truncating them to X.Y.Z, so a
publish from an untagged checkout can't ship a version that corresponds
to no release.
