+++ js/package.json
@@ -27,5 +27,5 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "version": "0.2.18"
+  "version": "0.0.0"
 }
